Move MenuItem into its own file
