+++ src/view/home/MenuView.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "view/BaseView.hpp"
+#include "view/home/MenuItem.hpp"
 
 #include <string>
 #include <vector>
@@ -10,30 +11,6 @@ struct tagMENU;
 
 namespace memo {
 namespace ui {
-
-class MenuItem
-{
-public:
-    MenuItem();
-    MenuItem(int iId, const std::string& iName, const std::string& iDescription="");
-    MenuItem(const MenuItem& iOther);
-    MenuItem(const MenuItem&& iOther);
-    MenuItem& operator=(const MenuItem& iOther);
-
-    int getId();
-    void setId(int iId);
-
-    const std::string& getName() const;
-    void setName(const std::string& iName);
-
-    const std::string& getDescription() const;
-    void setDescription(const std::string iDescription);
-
-private:
-    int id_;
-    std::string name_;
-    std::string description_;
-};
 
 class MenuView : public BaseView
 {
